Extract all metadata-related code to class without breaking any tests
